+++ ReleaseNotes.md
@@ -1,6 +1,16 @@
 <div style="text-align:right"><img src="https://raw.githubusercontent.com/gematik/gematik.github.io/master/Gematik_Logo_Flag_With_Background.png" width="250" height="47" alt="gematik GmbH Logo"/> <br/> </div> <br/>
 
 # Release notes
+## Release 1.1.1
+- added vex documents
+- updated spring-parent to version 4.1.12
+- updated base-image to 1.0.8
+- configured rabbitmq dynamic setting to allow the configuration of bulk.exchange for user svc-bulk-inbound-service
+- Updated minor/patch versions of dependencies
+- hardened connection to rabbitmq during roling update and unavailability of service
+- inqueue message handler: do not retry error message in case of bad smg request -> In this case batch will never finish
+- upload endpoint: bad request in case of an empty docId in header
+
 ## Release 1.1.0
 - updated MaxRamPercentage and RAM-Limits
 - updated jvm params and secret handling in helm charts